+++ src/main/java/no/uib/inf112/core/io/InputHandler.java
@@ -1,8 +1,10 @@
 package no.uib.inf112.core.io;
 
 import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Input;
 import com.badlogic.gdx.InputAdapter;
 import no.uib.inf112.core.RoboRally;
+import no.uib.inf112.core.player.Player;
 
 public class InputHandler extends InputAdapter {
