+++ src/main/java/no/uib/inf112/core/ui/CardContainer.java
@@ -3,8 +3,8 @@ package no.uib.inf112.core.ui;
 import no.uib.inf112.core.player.Card;
 import no.uib.inf112.core.player.Deck;
 import no.uib.inf112.core.player.Player;
-import no.uib.inf112.core.ui.cards.CardSlot;
-import no.uib.inf112.core.ui.cards.SlotType;
+import no.uib.inf112.core.ui.actors.cards.CardSlot;
+import no.uib.inf112.core.ui.actors.cards.SlotType;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Arrays;
