+++ src/main/java/no/uib/inf112/core/ui/actors/DisabledVisualizer.java
@@ -1,4 +1,4 @@
-package no.uib.inf112.core.ui;
+package no.uib.inf112.core.ui.actors;
 
 import com.badlogic.gdx.graphics.Color;
